Set up UserAccount screen.
